fix: price with float value

+++ src/Model/Price.php
@@ -4,6 +4,6 @@ namespace Scraper\ScraperZakeke\Model;
 
 class Price
 {
-    public ?int $value = null;
+    public ?float $value = null;
     public ?string $currency = null;
 }
